Support Ruby 3 keyword args in isolated methods
Ruby 3 no longer captures keyword arguments in `*args`, which meant that
keyword arguments got removed when the method was isolated.

This change works the same in Ruby 2.6, 2.7, and 3.0.

This should not be a breaking change, since all `details_message` procs
that only accept two arguments will get the keyword arguments added as a
hash to the end of the `args` parameter. If you implement a callback
that receives three arguments, then kwargs will be emitted as the third
argument instead.